[GCOS] Add support for paragraphs that do not terminate with a period
In fixed form programs, section and paragraph headers must start in
area A, and in the GCOS dialect this form is exploited to allow
paragraphs that do not terminate with a period in procedure divisions.
GNUCobol currently discards any information about the use of area A,
since every dialect it supports requires paragraphs to terminate with
a period.

To support the GCOS dialect, the suggested changes modify the
pre-processor to insert special markers `#area_a' before every word
that starts in area A.  This information is then exploited by the
parser to detect paragraph termination within procedure divisions.