[GCOS] Add support for sections and paragraphs that do not terminate with a period
In fixed form programs, section and paragraph headers must start in
area A, and in the GCOS dialect this form is exploited to allow
paragraphs that do not terminate with a period in procedure divisions.
Up to now, GNUCobol discarded any information about the use of area A,
since every supported dialect requires paragraphs to terminate with a
period.

To support the GCOS dialect, the suggested changes modify the
pre-processor to insert special markers `#area_a` before every word
that starts in area A when a new option `rely-on-area-a` is set.  The
parser can now make use of these tokens to detect section and
paragraph terminations within procedure divisions.